decouple scenario from commands and runners
