+++ src/cicadad/core/decorators.py
@@ -1,15 +1,15 @@
 from typing import Any, Callable, Union
 
-from cicadad.core.scenario import (
+from cicadad.core.types import (
     ConsoleMetricDisplays,
     MetricCollector,
-    Scenario,
     LoadModelFn,
     UserLoopFn,
     ResultAggregatorFn,
     ResultVerifierFn,
     OutputTransformerFn,
 )
+from cicadad.core.scenario import Scenario
 from cicadad.core.engine import Engine
 
 
@@ -44,7 +44,7 @@ def user_loop(user_loop_fn: UserLoopFn):
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "user_loop", user_loop_fn)
+        _set_scenario_attribute(fn, "user_loop", user_loop_fn)
 
         return fn
 
@@ -60,7 +60,7 @@ def users_per_container(users_per_container: int):
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "users_per_container", users_per_container)
+        _set_scenario_attribute(fn, "users_per_container", users_per_container)
 
         return fn
 
@@ -68,15 +68,14 @@ def users_per_container(users_per_container: int):
 
 
 def load_model(load_model_fn: LoadModelFn):
-    """Function to handle how scenario is run with regards to starting users and
-    administering work
+    """Handle how scenario is run with regards to starting users and administering work.
 
     Args:
         load_model_fn (LoadModelFn): User defined load model function
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "load_model", load_model_fn)
+        _set_scenario_attribute(fn, "load_model", load_model_fn)
 
         return fn
 
@@ -84,20 +83,20 @@ def load_model(load_model_fn: LoadModelFn):
 
 
 def dependency(dep: Scenario):
-    """Set a scenario as a dependency for this scenario to run
+    """Set a scenario as a dependency for this scenario to run.
 
     Args:
         dep (Scenario): Scenario this function is dependent on being successful
     """
 
     def wrapper(fn):
-        dependencies = get_scenario_attribute(fn, "dependencies")
+        dependencies = _get_scenario_attribute(fn, "dependencies")
         entry = [dep]
 
         if dependencies is None:
-            set_scenario_attribute(fn, "dependencies", entry)
+            _set_scenario_attribute(fn, "dependencies", entry)
         else:
-            set_scenario_attribute(fn, "dependencies", dependencies + entry)
+            _set_scenario_attribute(fn, "dependencies", dependencies + entry)
 
         return fn
 
@@ -105,15 +104,16 @@ def dependency(dep: Scenario):
 
 
 def result_aggregator(aggregator_fn: ResultAggregatorFn):
-    """Transform previous aggregate and list of results into an aggregated
-    single result. Called by load model function
+    """Transform previous aggregate and list of results into an aggregated single result.
+
+    Called by load model function.
 
     Args:
         aggregator_fn (ResultAggregatorFn): Aggregator function
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "result_aggregator", aggregator_fn)
+        _set_scenario_attribute(fn, "result_aggregator", aggregator_fn)
 
         return fn
 
@@ -121,15 +121,16 @@ def result_aggregator(aggregator_fn: ResultAggregatorFn):
 
 
 def result_verifier(verifier_fn: ResultVerifierFn):
-    """Create error messages for errors found in a list of results. Called by
-    load model function
+    """Create error messages for errors found in a list of results.
+
+    Called by load model function.
 
     Args:
         verifier_fn (ResultVerifierFn): Verifier function
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "result_verifier", verifier_fn)
+        _set_scenario_attribute(fn, "result_verifier", verifier_fn)
 
         return fn
 
@@ -137,15 +138,14 @@ def result_verifier(verifier_fn: ResultVerifierFn):
 
 
 def output_transformer(transformer_fn: OutputTransformerFn):
-    """Transform the aggregated result of the scenario after load model is
-    called
+    """Transform the aggregated result of the scenario after load model is called.
 
     Args:
         transformer_fn (OutputTransformerFn): Transformer function
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "output_transformer", transformer_fn)
+        _set_scenario_attribute(fn, "output_transformer", transformer_fn)
 
         return fn
 
@@ -160,13 +160,13 @@ def metrics_collector(collector: MetricCollector):
     """
 
     def wrapper(fn):
-        collectors = get_scenario_attribute(fn, "metric_collectors")
+        collectors = _get_scenario_attribute(fn, "metric_collectors")
         entry = [collector]
 
         if collectors is None:
-            set_scenario_attribute(fn, "metric_collectors", entry)
+            _set_scenario_attribute(fn, "metric_collectors", entry)
         else:
-            set_scenario_attribute(fn, "metric_collectors", collectors + entry)
+            _set_scenario_attribute(fn, "metric_collectors", collectors + entry)
 
         return fn
 
@@ -174,14 +174,14 @@ def metrics_collector(collector: MetricCollector):
 
 
 def console_metric_displays(displays: ConsoleMetricDisplays):
-    """Sets map of names to metric displays for scenario.
+    """Set map of names to metric displays for scenario.
 
     Args:
         displays (ConsoleMetricDisplays): Map of names to console metric display getters
     """
 
     def wrapper(fn):
-        set_scenario_attribute(fn, "console_metric_displays", displays)
+        _set_scenario_attribute(fn, "console_metric_displays", displays)
 
         return fn
 
@@ -189,21 +189,22 @@ def console_metric_displays(displays: ConsoleMetricDisplays):
 
 
 def tag(tag: str):
-    """Add a tag to a scenario. Tags allow for a test run to only cover certain
-    scenarios.
+    """Add a tag to a scenario.
+
+    Tags allow for a test run to only cover certain scenarios.
 
     Args:
         tag (str): Tag name to add to scenario
     """
 
     def wrapper(fn):
-        tags = get_scenario_attribute(fn, "tags")
+        tags = _get_scenario_attribute(fn, "tags")
         entry = [tag]
 
         if tags is None:
-            set_scenario_attribute(fn, "tags", entry)
+            _set_scenario_attribute(fn, "tags", entry)
         else:
-            set_scenario_attribute(fn, "tags", tags + entry)
+            _set_scenario_attribute(fn, "tags", tags + entry)
 
         return fn
 
@@ -220,7 +221,7 @@ def make_scenario(name: str, fn: Callable):
     return Scenario(name=name, fn=fn, **scenario_kwargs)
 
 
-def get_scenario_attribute(obj: Union[Scenario, Callable], name: str):
+def _get_scenario_attribute(obj: Union[Scenario, Callable], name: str):
     if hasattr(obj, name):
         return getattr(obj, name)
 
@@ -230,7 +231,7 @@ def get_scenario_attribute(obj: Union[Scenario, Callable], name: str):
     return None
 
 
-def set_scenario_attribute(obj: Union[Scenario, Callable], name: str, value: Any):
+def _set_scenario_attribute(obj: Union[Scenario, Callable], name: str, value: Any):
     if hasattr(obj, name):
         return setattr(obj, name, value)
